Started About page and better Events page
The events page has been made to look much better.
The code in this regard is still quite awful.
The about page has been started. Logistics about how the text should be
aligned must be considered.
There exists a bug where the facebook graph request is being called
every time the orientation changes. This needs to be fixed.
Not too attached to the border color for the events.