About Links, Events Borders, Gallery, File Cleanup
Links have been added to the social media section of the contact page.
Changed the color of the borders on the vents page.
Added images to the gallery, but they are not clickable.
Gallery images are different sizes. This must change
Layouts must be adjusted on the gallery page as they span off the
screen.
Cleaned up a few of the project files that weren't being used.