Action Bar Images and File Clean-up
Images have been added to the action bar.
Split the code for different fragments into different files to eliminate
the wall of text that was the main activity and reduce use of
unnecessary code.
Eliminated a bug where the gallery fragment would load as the first open
fragment as it was being loaded in an asynchronous task.
Removed a redundant file from the project.