Support Fragments, Social Buttons, Event Click Listeners
Changed the fragments across the app from app.Fragment to
app.support.v4.Fragment for better functionality with ViewPager and
better functionality overall.
Changed the social buttons from an ImageView next to a TextView to a
single Button. This required scaling down the social icon images, but
has made the code for this page smaller and easier to read/understand.
There are now click listeners associated with events. The code behind
this is not nearly done and quite sloppy/copied from others. More work
is needed.