Better Full Photos, Static Title, Better Loading, Removed Logging
Full photos now fit the screen better and there is a black background.
Changed the head layout title to UPAC. This only displays for viewpager
pages and should, in the future, dynamically display the full image name
or event name.
Removed most logging from the app in preparation for publishing.