Winona State UPAC 1.1.a
Changed the text sizes of the About page so the text is slightly large.
Changed how much of the screen the bottom bar takes up to fit it better
on tablets and bigger devices.
Fixed a bug where clicking the Events button right after opening the app
crashed the app.